// app/models/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ConnectionCursor: any;
  ConnectionLimitInt: any;
  Date: any;
  DateTime: any;
  Email: any;
  JSONObject: any;
};
// ...
export type CatalogBooleanFilter = {
  /** The name of the filter */
  name: CatalogBooleanFilterName;
  /** The filter value */
  value: Scalars['Boolean'];
};

/** The list of currently supported top level Catalog props on which catalog items can be filtered. */
export enum CatalogBooleanFilterName {
  /** isBackorder */
  IsBackorder = 'isBackorder',
  /** isDeleted */
  IsDeleted = 'isDeleted',
  /** isLowQuantity */
  IsLowQuantity = 'isLowQuantity',
  /** isSoldOut */
  IsSoldOut = 'isSoldOut',
  /** isVisible */
  IsVisible = 'isVisible',
}

/** Catalog items are combined to create a catalog. Each item can represent a different type of content. */
export type CatalogItem = {
  /** Item ID */
  _id: Scalars['ID'];
  /** Date and time at which the item was created */
  createdAt?: Maybe<Scalars['DateTime']>;
  /** Shop that owns the item */
  shop: Shop;
  /** Date and time at which the item was last updated */
  updatedAt?: Maybe<Scalars['DateTime']>;
};
// ...
export type CatalogItemConnection = {
  __typename?: 'CatalogItemConnection';
  /** The list of nodes that match the query, wrapped in an edge to provide a cursor string for each */
  edges?: Maybe<Array<Maybe<CatalogItemEdge>>>;
  /**
   * You can request the `nodes` directly to avoid the extra wrapping that `NodeEdge` has,
   * if you know you will not need to paginate the results.
   */
  nodes?: Maybe<Array<Maybe<CatalogItem>>>;
  /** Information to help a client request the next or previous page */
  pageInfo: PageInfo;
  /** The total number of nodes that match your query */
  totalCount: Scalars['Int'];
};
// ...
export type CatalogItemEdge = {
  __typename?: 'CatalogItemEdge';
  /** The cursor that represents this node in the paginated results */
  cursor: Scalars['ConnectionCursor'];
  /** The catalog item */
  node?: Maybe<CatalogItem>;
};
// ...
export enum CatalogItemSortByField {
  /** Sort by item ID */
  Id = '_id',
  /** Sort by date and time at which the item was created */
  CreatedAt = 'createdAt',
  /** Sort in the shop-defined order for the tag filter */
  Featured = 'featured',
  /** Sort by price */
  MinPrice = 'minPrice',
  /** Sort by date and time at which the item was last updated */
  UpdatedAt = 'updatedAt',
}
// ...
export type Deletable = {
  /**
   * If true, this object should be considered deleted. Soft deleted objects are not
   * returned in query results unless you explicitly ask for them.
   */
  isDeleted: Scalars['Boolean'];
};
// ...
export type Metafield = {
  __typename?: 'Metafield';
  /** Field description */
  description?: Maybe<Scalars['String']>;
  /** Field key */
  key?: Maybe<Scalars['String']>;
  /** Field namespace */
  namespace?: Maybe<Scalars['String']>;
  /** Field scope */
  scope?: Maybe<Scalars['String']>;
  /** Field value */
  value?: Maybe<Scalars['String']>;
  /** Field value type */
  valueType?: Maybe<Scalars['String']>;
};
// ...
export type Node = {
  /** The ID of the object */
  _id: Scalars['ID'];
};

/**
 * Objects implementing the NodeEdge interface will always have a node and a cursor
 * that represents that node for purposes of requesting paginated results.
 */
export type NodeEdge = {
  /** The cursor that represents this node in the paginated results */
  cursor: Scalars['ConnectionCursor'];
  /** The node itself */
  node?: Maybe<Node>;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['ConnectionCursor']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['ConnectionCursor']>;
};
// ...
export type QueryCatalogItemsArgs = {
  after?: InputMaybe<Scalars['ConnectionCursor']>;
  before?: InputMaybe<Scalars['ConnectionCursor']>;
  booleanFilters?: InputMaybe<Array<InputMaybe<CatalogBooleanFilter>>>;
  first?: InputMaybe<Scalars['ConnectionLimitInt']>;
  last?: InputMaybe<Scalars['ConnectionLimitInt']>;
  offset?: InputMaybe<Scalars['Int']>;
  searchQuery?: InputMaybe<Scalars['String']>;
  shopIds: Array<InputMaybe<Scalars['ID']>>;
  sortBy?: InputMaybe<CatalogItemSortByField>;
  sortByPriceCurrencyCode?: InputMaybe<Scalars['String']>;
  sortOrder?: InputMaybe<SortOrder>;
  tagIds?: InputMaybe<Array<InputMaybe<Scalars['ID']>>>;
};
// ...
export type QueryTagsArgs = {
  after?: InputMaybe<Scalars['ConnectionCursor']>;
  before?: InputMaybe<Scalars['ConnectionCursor']>;
  excludedTagIds?: InputMaybe<Array<InputMaybe<Scalars['ID']>>>;
  filter?: InputMaybe<Scalars['String']>;
  first?: InputMaybe<Scalars['ConnectionLimitInt']>;
  isTopLevel?: InputMaybe<Scalars['Boolean']>;
  last?: InputMaybe<Scalars['ConnectionLimitInt']>;
  offset?: InputMaybe<Scalars['Int']>;
  shopId: Scalars['ID'];
  shouldIncludeDeleted?: InputMaybe<Scalars['Boolean']>;
  shouldIncludeInvisible?: InputMaybe<Scalars['Boolean']>;
  sortBy?: InputMaybe<TagSortByField>;
  sortOrder?: InputMaybe<SortOrder>;
};
// ...
export type Shop = Node & {
  __typename?: 'Shop';
  /** The shop ID */
  _id: Scalars['ID'];
  /** An the shop's default address */
  addressBook?: Maybe<Array<Maybe<Address>>>;
  /** Whether to allow user to checkout without creating an account */
  allowGuestCheckout?: Maybe<Scalars['Boolean']>;
  /** The base unit of length */
  baseUOL?: Maybe<Scalars['String']>;
  /** The base unit of Measure */
  baseUOM?: Maybe<Scalars['String']>;
  /** URLs for various shop assets in various sizes */
  brandAssets?: Maybe<ShopBrandAssets>;
  /** The default shop currency */
  currency: Currency;
  /** The default navigation tree for this shop */
  defaultNavigationTree?: Maybe<NavigationTree>;
  /** The ID of the shop's default navigation tree */
  defaultNavigationTreeId?: Maybe<Scalars['String']>;
  /** Default parcel size for this shop */
  defaultParcelSize?: Maybe<ShopParcelSize>;
  /** Shop description */
  description?: Maybe<Scalars['String']>;
  /** The shop's default email record */
  emails?: Maybe<Array<Maybe<EmailRecord>>>;
  /** Returns a list of groups for this shop, as a Relay-compatible connection. */
  groups?: Maybe<GroupConnection>;
  /** Shop's keywords */
  keywords?: Maybe<Scalars['String']>;
  /** Shop default language */
  language: Scalars['String'];
  /** Shop name */
  name: Scalars['String'];
  /** Returns a list of roles for this shop, as a Relay-compatible connection. */
  roles?: Maybe<RoleConnection>;
  /** Returns URLs for shop logos */
  shopLogoUrls?: Maybe<ShopLogoUrls>;
  /** Shop's type */
  shopType?: Maybe<Scalars['String']>;
  /** Shop's slug */
  slug?: Maybe<Scalars['String']>;
  /** Returns URLs for various storefront routes */
  storefrontUrls?: Maybe<StorefrontUrls>;
  /** Returns a paged list of tags for this shop */
  tags?: Maybe<TagConnection>;
  /** Shop default timezone */
  timezone?: Maybe<Scalars['String']>;
  /** The shop's units of length */
  unitsOfLength?: Maybe<Array<Maybe<UnitOfLength>>>;
  /** The shop's units of measure */
  unitsOfMeasure?: Maybe<Array<Maybe<UnitOfMeasure>>>;
};
// ...
export enum SortOrder {
  /** ascending */
  Asc = 'asc',
  /** descending */
  Desc = 'desc',
}
// ...
export type Tag = Deletable &
  Node & {
    __typename?: 'Tag';
    /** The tag ID */
    _id: Scalars['ID'];
    /** The date and time at which this tag was created */
    createdAt: Scalars['DateTime'];
    /** A string of the title to be displayed on a Tag Listing Page */
    displayTitle?: Maybe<Scalars['String']>;
    /** A list of the IDs of top products in this tag */
    featuredProductIds?: Maybe<Array<Maybe<Scalars['ID']>>>;
    /** A string containing the hero image url for a Tag Listing Page */
    heroMediaUrl?: Maybe<Scalars['String']>;
    /**
     * If `true`, this object should be considered deleted. Soft deleted objects are not
     * returned in query results unless you explicitly ask for them.
     */
    isDeleted: Scalars['Boolean'];
    /** If `true`, this tag should be shown at the top level of the tag hierarchy */
    isTopLevel: Scalars['Boolean'];
    /** If `true`, this tag's Tag Listing Page should be visible to the public */
    isVisible: Scalars['Boolean'];
    /** Arbitrary additional metadata about this tag */
    metafields?: Maybe<Array<Maybe<Metafield>>>;
    /** The display name for the tag. This is unique within a given shop. */
    name: Scalars['String'];
    /** The tag's position relative to other tags at the same level of the tag hierarchy */
    position?: Maybe<Scalars['Int']>;
    /** The shop to which this tag belongs */
    shop: Shop;
    /** A unique URL-safe string representing this tag for links */
    slug?: Maybe<Scalars['String']>;
    /** A list of the IDs of tags that have this tag as their parent in the tag hierarchy, in the user-defined order */
    subTagIds: Array<Maybe<Scalars['ID']>>;
    /** A paged list of tags that have this tag as their parent in the tag hierarchy. Currently only three levels are supported. */
    subTags?: Maybe<TagConnection>;
    /** The date and time at which this tag was last updated */
    updatedAt: Scalars['DateTime'];
  };
// ...
export type TagConnection = {
  __typename?: 'TagConnection';
  /** The list of nodes that match the query, wrapped in an edge to provide a cursor string for each */
  edges?: Maybe<Array<Maybe<TagEdge>>>;
  /**
   * You can request the `nodes` directly to avoid the extra wrapping that `NodeEdge` has,
   * if you know you will not need to paginate the results.
   */
  nodes?: Maybe<Array<Maybe<Tag>>>;
  /** Information to help a client request the next or previous page */
  pageInfo: PageInfo;
  /** The total number of nodes that match your query */
  totalCount: Scalars['Int'];
};

/** A connection edge in which each node is a `Tag` object */
export type TagEdge = NodeEdge & {
  __typename?: 'TagEdge';
  /** The cursor that represents this node in the paginated results */
  cursor: Scalars['ConnectionCursor'];
  /** The tag */
  node?: Maybe<Tag>;
};
// ...
export enum TagSortByField {
  /** Tag ID */
  Id = '_id',
  /** Date and time the tag was created */
  CreatedAt = 'createdAt',
  /** Tag name */
  Name = 'name',
  /** Tag position */
  Position = 'position',
  /** Date and time the tag was last updated */
  UpdatedAt = 'updatedAt',
}